+++ lib/internal/bootstrap/pre_execution.js
@@ -27,6 +27,7 @@ function prepareMainThreadExecution(expandArgv1 = false) {
   // Patch the process object with legacy properties and normalizations
   patchProcessObject(expandArgv1);
   setupTraceCategoryState();
+  setupPerfHooks();
   setupInspectorHooks();
   setupWarningHandler();
 
@@ -220,6 +221,11 @@ function setupTraceCategoryState() {
   const { isTraceCategoryEnabled } = internalBinding('trace_events');
   const { toggleTraceCategoryState } = require('internal/process/per_thread');
   toggleTraceCategoryState(isTraceCategoryEnabled('node.async_hooks'));
+}
+
+function setupPerfHooks() {
+  require('internal/perf/performance').refreshTimeOrigin();
+  require('internal/perf/utils').refreshTimeOrigin();
 }
 
 function setupInspectorHooks() {
@@ -474,6 +480,7 @@ module.exports = {
   setupCoverageHooks,
   setupWarningHandler,
   setupDebugEnv,
+  setupPerfHooks,
   prepareMainThreadExecution,
   initializeDeprecations,
   initializeESMLoader,
