+++ lib/internal/http.js
@@ -9,7 +9,7 @@ const {
 
 const { setUnrefTimeout } = require('internal/timers');
 
-const { InternalPerformanceEntry } = require('internal/perf/perf');
+const { InternalPerformanceEntry } = require('internal/perf/performance_entry');
 
 const {
   enqueue,
