+++ lib/internal/main/worker_thread.js
@@ -18,6 +18,7 @@ const {
   setupInspectorHooks,
   setupWarningHandler,
   setupDebugEnv,
+  setupPerfHooks,
   initializeDeprecations,
   initializeWASI,
   initializeCJSLoader,
@@ -114,6 +115,7 @@ port.on('message', (message) => {
     } = message;
 
     setupTraceCategoryState();
+    setupPerfHooks();
     initializeReport();
     if (manifestSrc) {
       require('internal/process/policy').setup(manifestSrc, manifestURL);
