+++ lib/internal/perf/event_loop_utilization.js
@@ -2,7 +2,7 @@
 
 const nodeTiming = require('internal/perf/nodetiming');
 
-const { now } = require('internal/perf/perf');
+const { now } = require('internal/perf/utils');
 
 function eventLoopUtilization(util1, util2) {
   const ls = nodeTiming.loopStart;
