+++ lib/internal/perf/nodetiming.js
@@ -3,15 +3,14 @@
 const {
   ObjectDefineProperties,
   ObjectSetPrototypeOf,
-  SafeArrayIterator,
-  SafeSet,
 } = primordials;
 
+const { PerformanceEntry } = require('internal/perf/performance_entry');
+
 const {
-  PerformanceEntry,
-  kReadOnlyAttributes,
   now,
-} = require('internal/perf/perf');
+  getMilestoneTimestamp,
+} = require('internal/perf/utils');
 
 const {
   customInspectSymbol: kInspect,
@@ -29,25 +28,7 @@ const {
     NODE_PERFORMANCE_MILESTONE_ENVIRONMENT
   },
   loopIdleTime,
-  milestones,
-  timeOrigin,
 } = internalBinding('performance');
-
-function getMilestoneTimestamp(milestoneIdx) {
-  const ns = milestones[milestoneIdx];
-  if (ns === -1)
-    return ns;
-  return ns / 1e6 - timeOrigin;
-}
-
-const readOnlyAttributes = new SafeSet(new SafeArrayIterator([
-  'nodeStart',
-  'v8Start',
-  'environment',
-  'loopStart',
-  'loopExit',
-  'bootstrapComplete',
-]));
 
 class PerformanceNodeTiming {
   constructor() {
@@ -158,10 +139,6 @@ class PerformanceNodeTiming {
       loopExit: this.loopExit,
       idleTime: this.idleTime,
     };
-  }
-
-  static get [kReadOnlyAttributes]() {
-    return readOnlyAttributes;
   }
 }
