+++ lib/internal/perf/observe.js
@@ -31,7 +31,7 @@ const {
 const {
   InternalPerformanceEntry,
   isPerformanceEntry,
-} = require('internal/perf/perf');
+} = require('internal/perf/performance_entry');
 
 const {
   codes: {
@@ -174,11 +174,13 @@ class PerformanceObserverEntryList {
 }
 
 class PerformanceObserver {
-  [kBuffer] = [];
-  [kEntryTypes] = new SafeSet();
-  [kType] = undefined;
-
   constructor(callback) {
+    // TODO(joyeecheung): V8 snapshot does not support instance member
+    // initializers for now:
+    // https://bugs.chromium.org/p/v8/issues/detail?id=10704
+    this[kBuffer] = [];
+    this[kEntryTypes] = new SafeSet();
+    this[kType] = undefined;
     validateCallback(callback);
     this[kCallback] = callback;
   }
