+++ lib/internal/perf/performance_entry.js
@@ -7,10 +7,6 @@ const {
 } = primordials;
 
 const {
-  timeOrigin,
-} = internalBinding('performance');
-
-const {
   customInspectSymbol: kInspect,
 } = require('internal/util');
 
@@ -21,12 +17,6 @@ const kType = Symbol('kType');
 const kStart = Symbol('kStart');
 const kDuration = Symbol('kDuration');
 const kDetail = Symbol('kDetail');
-const kReadOnlyAttributes = Symbol('kReadOnlyAttributes');
-
-function now() {
-  const hr = process.hrtime();
-  return (hr[0] * 1000 + hr[1] / 1e6) - timeOrigin;
-}
 
 function isPerformanceEntry(obj) {
   return obj?.[kName] !== undefined;
@@ -88,7 +78,5 @@ ObjectSetPrototypeOf(
 module.exports = {
   InternalPerformanceEntry,
   PerformanceEntry,
-  kReadOnlyAttributes,
   isPerformanceEntry,
-  now,
 };
