+++ lib/internal/perf/timerify.js
@@ -9,10 +9,8 @@ const {
   Symbol,
 } = primordials;
 
-const {
-  InternalPerformanceEntry,
-  now,
-} = require('internal/perf/perf');
+const { InternalPerformanceEntry } = require('internal/perf/performance_entry');
+const { now } = require('internal/perf/utils');
 
 const {
   validateFunction,
