+++ lib/internal/perf/usertiming.js
@@ -3,16 +3,13 @@
 const {
   ObjectKeys,
   SafeMap,
+  SafeSet,
+  SafeArrayIterator,
 } = primordials;
 
-const {
-  InternalPerformanceEntry,
-  kReadOnlyAttributes,
-  now,
-} = require('internal/perf/perf');
-
+const { InternalPerformanceEntry } = require('internal/perf/performance_entry');
+const { now } = require('internal/perf/utils');
 const { enqueue } = require('internal/perf/observe');
-
 const nodeTiming = require('internal/perf/nodetiming');
 
 const {
@@ -31,8 +28,15 @@ const {
 } = require('internal/errors');
 
 const marks = new SafeMap();
-const nodeTimingReadOnlyAttributes =
-  nodeTiming.constructor[kReadOnlyAttributes];
+
+const nodeTimingReadOnlyAttributes = new SafeSet(new SafeArrayIterator([
+  'nodeStart',
+  'v8Start',
+  'environment',
+  'loopStart',
+  'loopExit',
+  'bootstrapComplete',
+]));
 
 function getMark(name) {
   if (name === undefined) return;
