+++ lib/internal/worker.js
@@ -28,7 +28,7 @@ const {
 const EventEmitter = require('events');
 const assert = require('internal/assert');
 const path = require('path');
-const { timeOrigin } = internalBinding('performance');
+const { now } = require('internal/perf/utils');
 
 const errorCodes = require('internal/errors').codes;
 const {
@@ -502,12 +502,6 @@ function eventLoopUtilization(util1, util2) {
   const active_delta = active - util1.active;
   const utilization = active_delta / (idle_delta + active_delta);
   return { idle: idle_delta, active: active_delta, utilization };
-}
-
-// Duplicate code from performance.now() so don't need to require perf_hooks.
-function now() {
-  const hr = process.hrtime();
-  return (hr[0] * 1000 + hr[1] / 1e6) - timeOrigin;
 }
 
 module.exports = {
