+++ test/parallel/test-bootstrap-modules.js
@@ -23,6 +23,7 @@ const expectedModules = new Set([
   'Internal Binding module_wrap',
   'Internal Binding native_module',
   'Internal Binding options',
+  'Internal Binding performance',
   'Internal Binding process_methods',
   'Internal Binding report',
   'Internal Binding serdes',
@@ -60,6 +61,7 @@ const expectedModules = new Set([
   'NativeModule internal/fs/rimraf',
   'NativeModule internal/fs/watchers',
   'NativeModule internal/heap_utils',
+  'NativeModule internal/histogram',
   'NativeModule internal/idna',
   'NativeModule internal/linkedlist',
   'NativeModule internal/modules/run_main',
@@ -77,6 +79,14 @@ const expectedModules = new Set([
   'NativeModule internal/modules/esm/translators',
   'NativeModule internal/process/esm_loader',
   'NativeModule internal/options',
+  'NativeModule internal/perf/event_loop_utilization',
+  'NativeModule internal/perf/nodetiming',
+  'NativeModule internal/perf/observe',
+  'NativeModule internal/perf/performance',
+  'NativeModule internal/perf/performance_entry',
+  'NativeModule internal/perf/timerify',
+  'NativeModule internal/perf/usertiming',
+  'NativeModule internal/perf/utils',
   'NativeModule internal/priority_queue',
   'NativeModule internal/process/execution',
   'NativeModule internal/process/per_thread',
